refactor: shared utils for json and durations
One parseJson, one clone and the three deliberate duration voices
(list chip, compact SLA wording, byline prose) move to utils.ts — the
same home desk/src/utils.ts gives them — replacing three parseJson
copies and three private duration implementations.
